Created new updated testbench
New testbench made that should make logical sense, hasn't been debugged yet, and needs 2 components still in the form of modports and clocking blocks which are important for our design. After those are added this will be put into vivado and hopefully start to be ran.

The final thing that needs to be created is verification logic. Right now this code will just create and run random packets, but we need to write simple verification logic that will check to see if our result is correct, so we won't have to sift through it ourselves